Significant message improvements
Added bad behaviour pings to tell off Squirrel when we haven't been doing WvW content for 30+ minutes. Improved PadCenter function to allow padding characters. Made and added in custom emojis for the borderlands. Removed the pillars on the headers to tidy them up. Isolated the joke footer logic so it can be used by all the messages.